fix: login menu item renderer
